add cross validation to classifier test
I have substancially changed how NBCAnalyzer works.  It uses proper
cross-validation (instead of random replicates, which has no
interpretable statistics).

In the course of doing this, I realized that pollution was hapenning in
my Naive Bayes Classifier -- the training set included the test set.
Correcting this obviously changed the results, actually in a good way.

I altered the plotting of classifier results -- both the
cross-comparison one and the longitudinal one.  In the cross-comparison
one, I added more pairwise comparisons, so now the comparisons are
exhaustive.  In both cross-comparison and longitudinal plots of
classifier ability, I added the 95% confidence interval (hard-coded
value :S) given a test set size of 50.  I need to double check this, and
make it not hard-coded.