add dynamic significance level for classifier plots
Added a couple functions to determine k_star and theta_star more easily,
and encorporated them into the plotter for the classifier-related plots.
Right now the caller has to specify $n$ and $\alpha$ in the plotting function
to get the significance line to show.  Ideally this would be embedded in
the data that the plotter reads, since this info is available to the
analysis function that creates the data, and because $n$ is specific to
the data produced for a given analysis.